requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
django-allauth==65.12.0
django-guardian==3.2.0
django-otp==1.6.1
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
